+++ DemoApp/src/main/java/com/neotreks/accuterra/mobile/demo/settings/DistanceFormatter.kt
@@ -3,5 +3,5 @@ package com.neotreks.accuterra.mobile.demo.settings
 import android.content.Context
 
 interface DistanceFormatter {
-    fun formatDistance(context: Context, distanceInMeters: Double): String
+    fun formatDistance(context: Context, distanceInMeters: Float): String
 }
