+++ DemoApp/src/main/java/com/neotreks/accuterra/mobile/demo/settings/DrivingTimeFormatter.kt
@@ -3,5 +3,5 @@ package com.neotreks.accuterra.mobile.demo.settings
 import android.content.Context
 
 interface DrivingTimeFormatter {
-    fun formatDrivingTime(context: Context, timeInSeconds: Long): String
+    fun formatDrivingTime(context: Context, timeInSeconds: Int): String
 }
