+++ DemoApp/src/main/java/com/neotreks/accuterra/mobile/demo/trip/online/OnlineTripStatsFragment.kt
@@ -53,7 +53,7 @@ class OnlineTripStatsFragment : OnlineTripFragment() {
         })
     }
 
-    private fun getDistance(formatter: DistanceFormatter, value: Double?) : String {
+    private fun getDistance(formatter: DistanceFormatter, value: Float?) : String {
         if (value == null) {
             return requireContext().getString(R.string.general_na)
         }
